backend/pyproject.toml
[project]
name = "meme-backend"
version = "0.1.0"
description = "Backend API service for meme generator - OpenTelemetry workshop"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.0",
    "pillow>=10.4.0",
    "django-cors-headers>=4.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--reuse-db --nomigrations --ds=settings"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false